extracted api calls to folder
created src/api folder to put api call functionality inside

separates the code more for clearer code
