+++ src/api/helpers.ts
@@ -1,0 +1,2 @@
+export const randomIndex = (list: any[]) =>
+  Math.floor(Math.random() * list.length);
